image_prediction/requirements.txt
matplotlib
nltk
numpy
Pillow
argparse
pandas
retrying
Cython
colormath
colorthief
colour
opencv-python
scipy
